backend/requirements.txt
fastapi
python-jose[cryptography]
httpx
pydantic
